Adding performance-focused data structures: ChampPack and PackedChampBooleanArray.
The purpose of these structs is to improve some of the worst performing hot areas in lolzone (along with
presumably many other applications that process large numbers of matches). The old map-based packing method
was not particularly fast or memory efficient, and it turns out that a large percentage of time in lolzone
is spent allocating, deallocating, and hashing for maps.

- ChampPack is a low-level data structure that provides a direct mapping between RiotID's and a new numerical
types, packedChampID's. A packedChampID is simply an int, but the new type exists for type checking purposes.
ChampPack's are responsible for keeping track of this mapping and quickly converting from one to the other, which
can be done in a single array lookup.

- PackedChampBooleanArray's are a higher-level structure built on top of a ChampPack, and abstract the notion of
packing altogether while providing an array-like interface. This change modifies the Match.Pack() process to use
these arrays instead of maps and has achieved a 5.2x speedup on go1.7.1, which is awesome.